src/third_party/WebKit/LayoutTests/http/tests/loading/sxg/sxg-fallback-with-fragment.html

<!DOCTYPE html>
<title>Fallback redirect of SignedHTTPExchange should preserve URL fragment</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  await waitUntilDidFinishLoadForFrame;

  const message = await openSXGInIframeAndWaitForMessage(t, 'resources/sxg-invalid-validity-url.sxg#fragment');
  assert_equals(message.location, 'https://127.0.0.1:8443/loading/sxg/resources/inner-url.html#fragment');
  assert_true(message.is_fallback);
}, 'Fallback redirect of SignedHTTPExchange should preserve URL fragment');
</script>
</body>
